refactor: extract Logseq behaviour to service
